Refactor: Taking out account_create from entry_page
