+++ members/static/members/sass/entry_page.scss
@@ -47,10 +47,3 @@
         }
     }
 }
-
-#signup-form {
-    border: 1px solid rgba(0, 0, 0, 0.125);
-    padding: 1.25rem;
-    background-clip: border-box;
-    border-radius: 0.25rem;
-}
