+++ members/tests/test_functional/test_admin_can_load.py
@@ -34,7 +34,7 @@ class SignUpTest(StaticLiveServerTestCase):
         self.browser.save_screenshot("test-screens/admin_load_test.png")
         self.browser.quit()
 
-    def test_entry_page(self):
+    def test_admin_page(self):
         # Loads the admin login page
         self.browser.get(f"{self.live_server_url}/admin")
         self.assertIn("admin", self.browser.title)
