+++ members/views/__init__.py
@@ -1,20 +1,21 @@
 # flake8: noqa  # ignored since it is being used in the files
-from members.views.FamilyDetails import FamilyDetails
+from members.views.AccountCreate import AccountCreate
+from members.views.Activities import Activities
+from members.views.ActivitySignup import ActivitySignup
+from members.views.AdminSignup import AdminSignup
 from members.views.ConfirmFamily import ConfirmFamily
+from members.views.DeclineInvitation import DeclineInvitation
+from members.views.DepartmentSignView import DepartmentSignView
+from members.views.EntryPage import EntryPage
+from members.views.FamilyDetails import FamilyDetails
+from members.views.Membership import Membership
 from members.views.PersonCreate import PersonCreate
 from members.views.PersonUpdate import PersonUpdate
-from members.views.WaitingListSetSubscription import WaitingListSetSubscription
-from members.views.DeclineInvitation import DeclineInvitation
-from members.views.ActivitySignup import ActivitySignup
+from members.views.QuickpayCallback import QuickpayCallback
+from members.views.SupportMembership import SupportMembership
 from members.views.UpdatePersonFromForm import UpdatePersonFromForm
-from members.views.EntryPage import EntryPage
+from members.views.WaitingListSetSubscription import WaitingListSetSubscription
+from members.views.departmentView import departmentView
+from members.views.paymentGatewayErrorView import paymentGatewayErrorView
 from members.views.userCreated import userCreated
 from members.views.volunteerSignup import volunteerSignup
-from members.views.QuickpayCallback import QuickpayCallback
-from members.views.DepartmentSignView import DepartmentSignView
-from members.views.paymentGatewayErrorView import paymentGatewayErrorView
-from members.views.departmentView import departmentView
-from members.views.Activities import Activities
-from members.views.Membership import Membership
-from members.views.SupportMembership import SupportMembership
-from members.views.AdminSignup import AdminSignup
